python/requirements.txt
aiohttp>=3.9
beautifulsoup4>=4.12
readability-lxml>=0.8
lxml>=5.0
spacy>=3.7
sentence-transformers>=2.2
numpy>=1.24
scikit-learn>=1.3
rake-nltk>=1.0
pydantic>=2.0
fastapi>=0.110
uvicorn>=0.27
click>=8.1
pyyaml>=6.0
